butler/backend/requirements.txt
sendgrid==6.10.0
psycopg2-binary==2.9.6
flask==2.0.2
pandas
requests
flask_cors
python-dotenv
